Created project files
created indexhtml and added canvas to generate google logo with zfont script which I had been wanting to attempt to play around with. Initially I wanted to try and get every letter to match the google logo colors but am having a hard time figuring out the best way to approach that. So I think im going to opt for a dark mode homepage layout just for fun and convenience.